requirements.txt
transformers 
transformers[sentencepiece]
datasets 
rouge_score 
py7zr 
pandas 
nltk 
tqdm 
PyYAML 
matplotlib 
torch 
notebook 
boto3 
mypy-boto3-s3 
python-box==6.0.2 
ensure==1.0.2 
fastapi 
uvicorn 
Jinja2 
#-e .